requirements.txt
# GUI 框架
PyQt6>=6.4.0

# 基础依赖
python-dotenv>=0.19.0  # 用于环境变量管理，如 API 密钥配置等

# 网络请求
requests>=2.28.0 